moving sections into another folder
